refactor(huggingface): Split conversion logic into multiple files.
